pyproject.toml
[project]
name = "podconvertation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==1.10.1",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.10.0",
    "av==15.1.0",
    "bitsandbytes>=0.49.2",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "ctranslate2==4.6.0",
    "decorator==4.4.2",
    "diffusers==0.35.1",
    "fastapi==0.116.2",
    "faster-whisper==1.2.0",
    "filelock==3.19.1",
    "flatbuffers==25.2.10",
    "fsspec==2025.9.0",
    "gtts==2.5.4",
    "h11==0.16.0",
    "hf-xet==1.1.10",
    "huggingface-hub==0.34.4",
    "humanfriendly==10.0",
    "idna==3.10",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "importlib-metadata==8.7.0",
    "jinja2==3.1.6",
    "lxml==6.0.1",
    "markupsafe==3.0.2",
    "moviepy==1.0.3",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numpy==2.3.3",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.3",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvtx-cu12==12.8.90",
    "omegaconf==2.3.0",
    "onnxruntime==1.22.1",
    "packaging==25.0",
    "pillow==11.3.0",
    "proglog==0.1.12",
    "protobuf==6.32.1",
    "psutil==7.0.0",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pydantic-settings>=2.13.1",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "python-pptx==1.0.2",
    "pyyaml==6.0.2",
    "regex==2025.9.1",
    "requests==2.32.5",
    "safetensors==0.6.2",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "starlette==0.48.0",
    "sympy==1.14.0",
    "tokenizers==0.22.0",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "tqdm==4.67.1",
    "transformers==4.56.1",
    "triton==3.4.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "xlsxwriter==3.2.8",
    "zipp==3.23.0",
]
